UIMA-2325 added functionality to remember and set initial  values of primitive variables.
git-svn-id: https://svn.apache.org/repos/asf/uima/sandbox/trunk/TextMarker@1229526 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-ep-textmarker-engine/src/main/java/org/apache/uima/textmarker/action/SetFeatureAction.java
@@ -127,7 +127,7 @@ public class SetFeatureAction extends AbstractTextMarkerAction {
             double v = numberExpr.getDoubleValue(element.getParent());
             annotationFS.setDoubleValue(featureByBaseName, v);
           } else if (range.equals(UIMAConstants.TYPE_FLOAT)) {
-            float v = (float) numberExpr.getDoubleValue(element.getParent());
+            float v = (float) numberExpr.getFloatValue(element.getParent());
             annotationFS.setFloatValue(featureByBaseName, v);
           } else if (range.equals(UIMAConstants.TYPE_BYTE)) {
             byte v = (byte) numberExpr.getIntegerValue(element.getParent());
